.history/requirements_20260403140817.txt
fastapi
uvicorn
python-dotenv
google-genai
Django>=4.2
mysqlclient
djangorestframework   # если хочешь делать API
python-dotenv